DUGK-68 Removed 'Components.tsx' and renamed '...Data.ts' (#47)
