--- client/src/route/library/Components.tsx
@@ -1,30 +1,0 @@
-/*
-src: https://mui.com/material-ui/react-tabs/
-*/
-import * as React from 'react';
-import TabComponent from 'components/tab/TabComponent';
-import { TabData } from 'components/tab/subcomponents/TabRender';
-import Iframe from 'components/Iframe';
-import tabs from './ComponentsData';
-
-const jupyterURL = window.env.REACT_APP_URL_LIB;
-
-const tabsData: TabData[] = tabs.map((tab) => ({
-  label: tab.label,
-  body: (
-    <>
-      {tab.body}
-      <Iframe
-        title={`JupyterLight-Demo-${tab.label}`}
-        url={`${jupyterURL}`}
-        fullsize
-      />
-    </>
-  ),
-}));
-
-function LibComponents() {
-  return <TabComponent tabs={tabsData} />;
-}
-
-export default LibComponents;
